backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
openai
pydantic
openai-whisper
pyttsx3